pyproject.toml
[tool.poetry]
name = "staking-bot-template"
version = "0.1.0"
description = ""
authors = ["Aloe Capital LLC"]

[tool.poetry.dependencies]
python = "^3.9"
web3 = "^5.20.0"
python-dotenv = "^0.17.1"
autopep8 = "^1.5.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
